Passthrough the opaque Js.Promise.error to handler
Currently, Vow requires that you ignore any error caught by
`Js.Promise.catch`.

Instead of this library deciding that the error handler cannot handle
the opaque error, it should pass it into the handler and let that
decide if it is able to parse what the error is and handle it if it can.